Extract CLI TUI agent runtime mixin
