+++ kolega_code/cli/tests/test_app_image_vision_warning.py
@@ -14,6 +14,8 @@ import base64
 
 import pytest
 
+from kolega_code.cli.tui import agent_runtime as agent_runtime_module
+
 from kolega_code.cli.config import build_agent_config, config_summary
 from kolega_code.cli.session_store import SessionStore
 
@@ -58,7 +60,6 @@ def _make_app(tmp_path, monkeypatch):
     """Build a KolegaCodeApp with a no-op FakeCoderAgent; agent set by the caller."""
     pytest.importorskip("textual")
 
-    from kolega_code.cli import app as app_module
     from kolega_code.cli.app import KolegaCodeApp
 
     class FakeCoderAgent:
@@ -80,7 +81,7 @@ def _make_app(tmp_path, monkeypatch):
         async def cleanup(self):
             return None
 
-    monkeypatch.setattr(app_module, "CoderAgent", FakeCoderAgent)
+    monkeypatch.setattr(agent_runtime_module, "CoderAgent", FakeCoderAgent)
 
     project = tmp_path / "project"
     project.mkdir()
